Prevent duplicated test class
